# expression: implement vectorized evaluation for builtin field(Real/Int/String)