--- a/expression/builtin.go
+++ b/expression/builtin.go
@@ -18,6 +18,7 @@
 //go:generate go run generator/control_vec.go
 //go:generate go run generator/time_vec.go
 //go:generate go run generator/compare_vec.go
+//go:generate go run generator/string_vec.go
 
 package expression