--- a/expression/generator/string_vec.go
+++ b/expression/generator/string_vec.go
@@ -76,16 +76,18 @@ func (b *builtinField{{ .TypeName }}Sig) vecEvalInt(input *chunk.Chunk, result *
 {{ end }}
 	result.ResizeInt64(n, false)
 	i64s := result.Int64s()
+	for i := 0; i < n; i++ {
+		i64s[i] = 0
+	}
 	for i := 1; i < len(b.args); i++ {
 		if err := b.args[i].VecEval{{ .TypeName }}(b.ctx, input, buf1); err != nil {
 			return err
 		}
-		buf1.MergeNulls(buf0)
 {{ if .Fixed }}
 		arg1 := buf1.{{ .TypeNameInColumn }}s()
 {{ end }}
 		for j := 0; j < n; j++ {
-			if buf1.IsNull(j) || i64s[j] > 0 {
+			if buf0.IsNull(j) || buf1.IsNull(j) || i64s[j] > 0 {
 				continue
 			}
 {{ if .Fixed }}