… expression/builtin_string_vec_generated_test.go