<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
use go generate to implement vectorized evaluation for builtin field(Real/Int/String)
[12058](https://github.com/pingcap/tidb/issues/12058)

### What is changed and how it works?
```
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedGeneratedBuiltinStringFunc/builtinFieldIntSig-VecBuiltinFunc-4         	   26844	     56042 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedGeneratedBuiltinStringFunc/builtinFieldIntSig-NonVecBuiltinFunc-4      	   10000	    149427 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedGeneratedBuiltinStringFunc/builtinFieldRealSig-VecBuiltinFunc-4        	   73128	     15201 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedGeneratedBuiltinStringFunc/builtinFieldRealSig-NonVecBuiltinFunc-4     	   21536	     49933 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedGeneratedBuiltinStringFunc/builtinFieldStringSig-VecBuiltinFunc-4      	   41544	     32878 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedGeneratedBuiltinStringFunc/builtinFieldStringSig-NonVecBuiltinFunc-4   	   15165	     76603 ns/op	       0 B/op	       0 allocs/op
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test
 - Integration test